refactor: split command palette dialog sources
